Capsuled, Updated README.md, Updated main().
